fix: bug in fragment spread counter in cache manager

+++ packages/cache-manager/src/helpers/filterFragmentDefinitions.ts
@@ -34,6 +34,10 @@ export default (
     return deleted > 0 && deleted === total ? [...names, key] : names;
   }, []);
 
+  if (!definitionsToDelete.length) {
+    return ast;
+  }
+
   return deleteFragmentDefinitions(ast, {
     include: definitionsToDelete,
   });
